Move routes into their own space
